TAx and Partner work

+++ DMS/Models/PartnerClasses/BindTaxData.cs
@@ -8,7 +8,7 @@ namespace DMS.Models.PartnerClasses
 {
     public class BindTaxData
     {
-        public Int32 TaxSetupInfoId { get; set; }
+        public Int64 TaxSetupInfoId { get; set; }
         public string FieldName { get; set; }
     }
 }
